[#183876777] Bug fix: Point selection in graphs should not append unless Shift key is down
* Renamed graph-hooks.tx to use-plot.ts
* Code review changes
