+++ v3/src/components/graph/components/casedots.tsx
@@ -4,7 +4,7 @@ import React, {memo, useCallback, useEffect, useRef, useState} from "react"
 import {pointRadiusSelectionAddend, transitionDuration} from "../graphing-types"
 import { ICase } from "../../../models/data/data-set"
 import { isAddCasesAction } from "../../../models/data/data-set-actions"
-import {useDragHandlers, usePlotResponders} from "../hooks/graph-hooks"
+import {useDragHandlers, usePlotResponders} from "../hooks/use-plot"
 import {useDataConfigurationContext} from "../hooks/use-data-configuration-context"
 import {useDataSetContext} from "../../../hooks/use-data-set-context"
 import {useInstanceIdContext} from "../../../hooks/use-instance-id-context"
