+++ v3/src/components/graph/hooks/use-plot.ts
@@ -1,16 +1,12 @@
-import {select} from "d3"
 import React, {useCallback, useEffect, useRef} from "react"
 import {reaction} from "mobx"
 import {onAction} from "mobx-state-tree"
-import {tip as d3tip} from "d3-v6-tip"
 import {isSelectionAction, isSetCaseValuesAction} from "../../../models/data/data-set-actions"
 import {INumericAxisModel} from "../models/axis-model"
 import {GraphLayout} from "../models/graph-layout"
 import {useCurrent} from "../../../hooks/use-current"
 import {IGraphModel} from "../models/graph-model"
-import {transitionDuration} from "../graphing-types"
-import {IDataSet} from "../../../models/data/data-set"
-import {getPointTipText, matchCirclesToData} from "../utilities/graph-utils"
+import {matchCirclesToData} from "../utilities/graph-utils"
 import {useInstanceIdContext} from "../../../hooks/use-instance-id-context"
 
 interface IDragHandlers {
@@ -149,53 +145,4 @@ export const usePlotResponders = (props: IPlotResponderProps) => {
     return () => disposer()
   }, [dataset, enableAnimation, graphModel, callRefreshPointPositions, dotsRef, instanceId])
 
-}
-
-const dataTip = d3tip().attr('class', 'graph-d3-tip')/*.attr('opacity', 0.8)*/
-  .attr('data-testid', 'graph-point-data-tip')
-  .html((d: string) => {
-    return "<p>" + d + "</p>"
-  })
-
-export const useDataTips = (dotsRef: React.RefObject<SVGSVGElement>,
-                            dataset: IDataSet | undefined, graphModel: IGraphModel) => {
-  const hoverPointRadius = graphModel.getPointRadius('hover-drag'),
-    pointRadius = graphModel.getPointRadius(),
-    selectedPointRadius = graphModel.getPointRadius('select'),
-    attrIDs = graphModel.config.uniqueTipAttributes
-
-  useEffect(() => {
-
-    function okToTransition(target: any) {
-      return target.node()?.nodeName === 'circle' && dataset && /*!active(target.node()) &&*/
-        !target.property('isDragging')
-    }
-
-    function showDataTip(event: MouseEvent) {
-      const target = select(event.target as SVGSVGElement)
-      if (okToTransition(target)) {
-        target.transition().duration(transitionDuration).attr('r', hoverPointRadius)
-        const [, caseID] = target.property('id').split("_"),
-          tipText = getPointTipText(caseID, attrIDs, dataset)
-        tipText !== '' && dataTip.show(tipText, event.target)
-      }
-    }
-
-    function hideDataTip(event: MouseEvent) {
-      const target = select(event.target as SVGSVGElement)
-      dataTip.hide()
-      if (okToTransition(target)) {
-        const [, caseID] = select(event.target as SVGSVGElement).property('id').split("_"),
-          isSelected = dataset?.isCaseSelected(caseID)
-        select(event.target as SVGSVGElement)
-          .transition().duration(transitionDuration)
-          .attr('r', isSelected ? selectedPointRadius : pointRadius)
-      }
-    }
-
-    select(dotsRef.current)
-      .on('mouseover', showDataTip)
-      .on('mouseout', hideDataTip)
-      .call(dataTip)
-  }, [dotsRef, dataset, attrIDs, hoverPointRadius, pointRadius, selectedPointRadius])
 }
